tokens: add ability to cherry-pick tokens from the WMF
Adds script(s) to extract a `tokens.json`, as compatible with
style-dictionary, from wikimedia-ui-base and adds those tokens as the
"base truth" to the configuration of style-dictionary.

`global.json` showcases the reuse of WMF's base tokens, e.g.
"wmui-color-base0" as "wmf-wmui-color-base0.value". I.e. names are
identical to what can be seen in [0] (ensure to pick the release
corresponding to the version as can be seen in our package.json) but get
prefixed by "wmf-".

[0]: https://github.com/wikimedia/wikimedia-ui-base/blob/master/wikimedia-ui-base.less