root: run lerna bootstrap upon npm install
This ensures that package-level deps (and all of lerna's magic) are
installed if npm is used to install (or ci) via npm.

This way around the setup is more in line with most of our other
projects (npm as the entrypoint) and it is also possible to use lerna
(after npm installation) without having lerna installed globally.