tokens: set global tokens
Uses a mixture of local values and values cherry-picked from the WMF.

We are trying to reuse as much as possible but ensure a structure we
deem consistent and add local values which we deem missing. We will
recommend the delta for upstreaming.

Technically this is the state as of [0] - the history did not contain
much information in the commit messages which made carrying it over not
worth it.

[0]: https://github.com/wmde/wikit-css/pull/22/commits/4387d23