Filter out inherited variables from output files
In our design system we only want to use tokens defined within it. We
pick WMF variables into our global files, but do not want to use them in
alias and component-level tokens. Most certainly we do not want them to
be used by consuming applications or showcase them in storybook for
direct consumption.