docs: illustrate cursors
This
* shows the cursor tokens and values as a table
* illustrates the cursor upon hover of the value column
* renders ids and (self) links for the cursor tokens (not entirely sure where this is going yet)
* shows a lack of "source" token(s) as dash

Technically, this uses the token name to index the tr loop (key) - the
property is required by mdx, I assume to ensure that it can avoid
re-rendering things should the list change dyamically (like in Vue), and
the name is unique and stable by definition so lends itself better for
this use than the, by definition dynamic, index of the array.

We don't have that many representatives but there is a pattern emerging
e.g. between the cursors and the opacity story. Will look into how this
can be abstracted in a follow-up.

Bug: T256861