Add a GitHub Action for Chromatic
Use npm script for the chromatic command to build storybook
THe github action is calling that script. We are not using
the official chromatic github action because it's currently
not possible to run it from a  monorepo
https://github.com/chromaui/action/issues/30

A "fake" component story is added because Chromatic does not detect
mdx files (without a component Story in them). If the storybook
does not have stories, the Chromatic CLI fails to verify the storybook
even though it builds it.