docs: infer presenter from token name
This allows to show tokens without deciding for one specific way of
presenting all of them (e.g. in a collaction of many different ones).

This leaves the color visualization in a "broken" state - the component
to visualize them so far has the display of token name and subtitle
backed into it and its "swatch" can not be used in isolation.

UX has called out the (subtle) discrepancy between the way we show
color's name + referenced tokens and other tokens' name + referenced
already, so we need to consolidate this one way or another.

Bug: T256024