Use custom kebab case function for generating token names
This overrides style-dictionary's own kebab case function to prevent it
from hyphenating parts of the path that contain numbers. For example,
style-dictionary produces the name `font-weight-style-h-1` from path
`[ 'font', 'weight', 'style', 'h1' ]`, but we want it to be
`font-weight-style-h1`.

This may become redundant should style-dictionary's own kebab case
function change according to what is discussed in
https://github.com/amzn/style-dictionary/issues/300

Bug: T257805