docs, vue-components: compose the storybooks
Build vue-components storybook into a subdirectory of docs and teach
that to use it as a ref for storybook "composition"[0].

We want the versions of our packages to be always aligned, the same goes
for their storybooks. As keeping those versions aligned when hosting the
storybooks on different domains is a challenge we instead host the
component lib(s) as a subfolder of the design system. This restricts the
compose functionality to "static" storybook builds (our dev container
does not hold the dependencies needed to build the `stories.json` for
"composition", and the storybook dev server obviously does not build
into a 3rd party folder by default) but devs have access to both
storybooks independently and will probably also prefer to use them that
way. Could be added later if somebody sees a need for it, however.

Also aligns the github workflow for branch build and master build to
ensure that root level npm script build:storybook always gets called
explicitly (which does all the necessary steps in the right order) and
only then the publishing happens (either to chromatic or github pages).

Chromatic will not pick up the sub-folder and consequently the review
workflow will not be available as it simply will not know the
vue-components' storybook even existing, if it does not get uploaded
with its own project key. This could be a way forward _should_ we ever
want UI review for components.

[0]: https://github.com/storybookjs/storybook/blob/3bc4666/docs/src/pages/configurations/composition/index.md

Bug: https://phabricator.wikimedia.org/T257824
Bug: https://phabricator.wikimedia.org/T258494