Make it possible to use logical properties
The lib has to support RTL languages.
CSS Logical Properties[0] offers the ability to control the layout
through logical, rather than physical, direction and dimension mappings.
They are supported in all major browsers but IE11[1].
That's why this patch introduces two postcss plugins which transform
1)css logical props and 2) the :dir selector[2]

This should be sufficient for devs to write css that supports LTR and
RTL languages.

https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Logical_Properties
https://caniuse.com/#search=logical%20properties
https://caniuse.com/#search=%3Adir