browser tests: make the story iframe selector generic
In CI we run our browser tests against a storybook that is a composition
of the storybooks contained in the docs/ and vue-components/ folders,
whereas locally they're only ran against the vue-components storybook.
The iframe containing the component story has a different id in both
cases which makes it difficult to select the right frame for the test.
This commit adjusts the `openComponentStory` command to always pick the
right frame.