vue-components: consistently apply button transitions
This changes the button transitions away from being state-specific, and
instead applies them equally for all state changes. They can still be
overridden in the respective state selectors if desired.

This means that the transitions show not only when transitioning from
the default state to the hover, active, or focus state, but also when
going back to the default state.