use phabricator as a bug tracker for all packages
The title should maybe read "for all published packages and the root
project" instead but it's cumbersome.

A short research did not yield any information if just putting it in the
root package.json would somehow have lerna take care of distributing it
to the packages but I wanted to be sure the information is part of the
published artifacts.