vue-components: plain storybook extract command
At the time of writing, storybook 6 was not released yet and running the
`extract` command on a storybook 5 (the npx default) caused an error.

Not sure if pinning to the now-current storybook version would be
better, but it's not how the storybook people document it[1], and will
cause future update effort (or inconsistency) while it is unlikely the
command will ever change without us having to touch the line anyways.

[1]: https://storybook.js.org/docs/vue/workflows/storybook-composition#improve-your-storybook-composition