Create Lookup component
The lookup component is a text input field that provides matching
selectable suggestions as a user types into it. In the context of
Wikidata, they can be used as Item and Property selectors, for example.

This contains the initial implementation of the visual and interactive
aspects of the Lookup component and its menu. The lookup behavior will
be implemented in a separate step.

Bug: T265008