Add Message component
Based on the Figma spec outlined in the ticket

Component level tokens are named semantically, from the design point of view.

The icon size is set to large, so it will be 20px with the default font
size as described by the specs.

The storybook is using the same explanatory copy used in the WMF's Style
guide

Using a slot instead of a prop of the message allows us to use markup
like bold and links in the message.

Bug: T265581

Co-Authored-by: Sarai Sanchez <sarai.sanchez@wikimedia.de>
Co-Authored-by: Michael Große <michael.grosse@wikimedia.de>