TextInput: use <label> for the label
Wrapping the label around the input element became questionable as soo
as we also included the validation message in the TextInput component.
We originally went with the label as the root element of the component
to work around having to generate an id for it.

According to the WCAG having the `for`-attribute is also beneficial for
screenreaders:
https://www.w3.org/TR/2008/WD-WCAG20-TECHS-20080430/H44.html