docs: list/visualize component-level tokens
Gets the tokens from the tokens' module export, filters it to
component level and groups them.

The resulting stories are a function of the existing components (i.e.
dynamic) which makes things much harder than one would think it needs to
be. Find a way to dynamically create navigation entries, use our custom
components in the mark up.

Bug: https://phabricator.wikimedia.org/T260425