vue-components: check components' wikit class on required props
The test was added to ensure that all components exhibit the class
"wikit" in their root node in order to be able to have a selector for
styles (resets).

At the time we missed that components may have required props which will
have to be specified to render (mount) the component.

This allows for a quick way to specify those props per component.
However, as this an endless job, I think we may be better off swapping
this jest test for something based on static analysis (e.g. inspired by
[0]). Granted, this can not do the full complexity of a vue component
justice, but we probably don't want dynamic behavior in the wikit class
anyway.

[0]: https://github.com/vuejs/eslint-plugin-vue/blob/80b8983/lib/rules/static-class-names-order.js