Keyboard navigation for selecting an item (#241)
* Fix encapsulation of test helper function

The name of the method implies that it returns a wrapper with an
expanded menu. However, for that to happen, vue needs to rerender (i.e.
a "tick" needs to pass). This should be done inside the method, not in
every test that calls it.

* Keyboard navigation for selecting an item

The input component should stay focused upon menu selection,
regardless of selection by keyboard or mouse, so the blurring on
selection was removed.

Bug: T265884

* Remove hasActiveElement property

There is no requirment to keep the active style for
keyboard and mouse selection the same. The active style
should only be applied upon mouse selection.

Co-authored-by: Michael Große <mic.grosse@googlemail.com>