Rename Button prop 'type' to 'flavor'
The Button component is simply a <button> HTML element
which means the native attribute 'type' cannot be set on this
component due to the component already having a prop named 'type'.
Renaming the prop to 'flavor' allows the consuming application to set
the type of Button to 'submit', 'reset' etc. while getting the desired
visual type of button e.g. 'primaryProgressive'.