Move keyboard navigation logic to LookupMenu
Having this logic inside the LookupMenu, makes it much easier to share
between consumers of that component. Also, since this is about keyboard
navigation of the LookupMenu, this seems to be the right place for it.

Since we are using a ref on the LookupMenu, it is important that this
menu always exists. This is achieved with using v-show instead of vi-if
on it. That has the consequence that the tests need to change and check
if the menu is visible instead of whether it exists.

This is done in preparation for extracting the Dropdown.

The tests are tricky insofar that on the one hand the behavior is now on
the LookupMenu component and should be tested there. On the other hand,
the behavior we actually care about (i.e. our "public interface") is the
Lookup component and the relevant question is whether it works correctly
together with its internally used components.

Sources for the type of the $ref:
* https://github.com/vuejs/vue-class-component/issues/94
* https://github.com/bluecanvas/vue-typed-refs

Bug: T266545