Make bold titles in LookupMenu optional
This is done in preparation for remaking the LookupMenu in to a more
general OptionsMenu that is then to also be used for the Dropdown. And
the Dropdown has Options who's label is not bold.

Two options and controls are added to the story to showcase this.

The two possible font weights of menu item labels (regular and bold) are
now represented by two different tokens.

Co-Authored-by: Sarai Sanchez <sarai.sanchez@wikimedia.de>