Add Dropdown component
The <label> does not seem to natively focus non-form elements even if
they have a tabindex >= 0 (at least on Firefox), so this is done
explicitly via a click event.

The SVG of the small arrow is added inline for now. This allows us to
style it with CSS, which wouldn't be possible if we were to use it via
background-image. It can be extract if and when we need to use it
somewhere else as well. It was not added to the Icon component as that
component explicitly sets a color instead of allowing the usage of the
current text color.

The scroll events are ignored as a Dropdown always has all its possible
options available.

Since the menu is always in the DOM due to the v-show, its
`mounted`-hook actually fires while it is still hidden and so cannot
update the height correctly. The watch on the `menuItems` never fires
because those items never change. Therefore, we have to trigger the
resize manually from the parent component after the Menu has rendered.

The validation message will be added in another commit.

Bug: T266545