vue-components: check components' wikit class via eslint
Use static analysis of the component code to make sure that all
components carry a "wikit" class on their root node so that our styles
and style resets can work properly (see #117).

The pre-existing test-driven approach had the disadvantage that one
needed to maintain the list of required props of all components for them
to be mountable.

Bug: https://phabricator.wikimedia.org/T267741