vue-components: Input: add classes in template
Add the classes to the Input component inside the template as
eslint-plugin-vue-root-class is not able to pick them up otherwise.

First I only moved the static classes up but then realized that, with
the possibilities of class binding[0] in vue, moving even the dynamic
part does not look that bad either.

[0]: https://vuejs.org/v2/guide/class-and-style.html