Only react to keydown events
In accordance with the official examples, it seems consistent to always
react to keydown events.
See
https://www.w3.org/TR/wai-aria-practices-1.1/examples/listbox/listbox-collapsible.html
for an example.

This has the positive side-effect that opening the menu on enter for the
upcoming dropdown will not cause problems where the menu is opened on
keydown and closed on the keyup of the same keypress.