Fix hovering the correct item after reopening the menu (#302)
* Fix hovering the correct item after reopening the menu

By waiting till after the next tick, we can ensure that also the
selectedItemIndex prop was updated, if applicable. If no item is
selected, then that index should be -1 just as expected.

* Only react to keydown events

In accordance with the official examples, it seems consistent to always
react to keydown events.
See
https://www.w3.org/TR/wai-aria-practices-1.1/examples/listbox/listbox-collapsible.html
for an example.

This has the positive side-effect that opening the menu on enter for the
upcoming dropdown will not cause problems where the menu is opened on
keydown and closed on the keyup of the same keypress.