Add Error and Warning state to Dropdown
The component refs are only properly typable inside the setup method,
however, we have to also use them there for TypeScript to understand the
type. This is similar to what has been done for the lookup in the
previous commit, albeit for a much simple method using the ref.

The PropType is removed from the prop definition as it seems to be
completely ignored. However, the menuItem prop must be defined in the
signature of the setup function even if it is not used inside the setup
function to receive typing in the computed methods added via the usual
options API.

Bug: T266546