Extract validation message logic with composition (#305)
With these being the first meaningful case of code-reuse, it is a good
moment to introduce the vue3 composition api via a plugin to this vue2
library.

To get the type inference for the setup option, we have to use the
defineComponent function for creating our components. This has the
side-effect that we have to rethink how to type the menu ref. The
solution was to move this into the setup function as well because there
we can use the vue3 way of accessing and _typing_ refs.

Bug: T266546