Update button tokens (#321)
* Create missing alias tokens

This adds the alias tokens that define the color value of the border and the background of disabled quiet buttons.

* DRY quiet alias tokens

* restructure button tokens

This organizes tokens per button variant and flavor.

* add icon-only button tokens

* adding icon spacing token

This defines the required spacing that should exist between the button's icon and its label

* update Button.vue variables

the variables used to style the Button.vue are replaced by the new ones derived from the new component tokens file

* Update tokens/properties/components/Button.json

* group global button icon tokens

* avoid renaming normal variant for now

The variant names "normal" and "framed" are used interchangeably in the documentation. "Framed" was introduced here to match the naming applied to the WVUI button. But I believe that this change belongs in a different PR.