Move global afterEach saucelabs logic into a custom command
The global afterEach function runs after each test *suite* which
is not ideal when we want to send info for each test that was executed.

The same logic is moved to a more appropriate place - a custom command.
This means every test suite needs to define an afterEach hook where the
custom command is called before ending the sesison.

This improvement fixes the problems with running browser tests locally.

Bug: T268975, T266343