Enforce Icon alignment responsibilities
Components should only handle their inner positioning and alignment, not
how they are positioned inside their parent. Consequently,
vertical-align: middle is removed from the Icon component.

As a consequence of that the inner alignment of the component that uses the
Icon, i.e. ValdationMessage, is adjusted.
Also adjusted is the Button story, as it now closer resembles what needs
to be done in the application: positioning the Icon inside the Button as
the Button cannot know what it will get inside its slot.