Use more semantic props
Renaming the value attribute to nativeValue just to have a value prop
pointing to the checked attribute is not very intuitive. These new props
are aligned to usage with the .sync modifier in vue2 and the new v-model
in vue3.

The basic story is slightly adjusted to showcase the use of the sync
modifier.