Remove the need for a Dockerfile & fix lerna usage
Tarrow started down the route of removing the need to
build a custom docker image in this repo, and I have taken
over the branch with this change fully removing the Dockerfile.

Change summary:
 - Remove Dockerfile, just using the base image now
 - Make all lerna commands use lerna from node_modules rather
   than any globally installed version.
 - Rework readme to just specify npm commands (not via docker-compose)
 - Add lerna command to package.json
 - Stop nightwatch from starting chromedriver when it is not needed

README moving away from docker-compose:
In other WMDE repos we moved away from including READMEs stating how
to run tooling via docker. Instead just using the much shorter, easier
to read base commands.
It would then be up to an individual to run npm, or other tooling in
a docker container if they want to for their development setup.

lerna usage:
The Dockerfile loaded lerna globally, and this was then inconsistently
used throughout various npm commands and README suggestions.
When reviewing other projects that use lerna they all seem to specify
the lerna version in package.json and just use the version installed
there, so I went for that approach.

nightwatch chromedriver:
Before this change nightwatch would start a chromedriver instance in
all situations, even when not used. It is not use for most situations
such as docker or saucelabs.
For now I have only set this to not start for the docker command.
This was needed as having chromedriver start means extra deps are needed
in the docker container.