refactor: extract core of LookupComponent (#379)
We'll need that core in the QuantityInput component. However, in that
component we have no use for the Label and ValidationMessage
subcomponents of the Lookup. Also, we need to be able to set attributes
directly on the input, to be able to set its id value from the outside
so that we can point the QuantityInput label to it.