ci: run on pull requests (#439)
Run CI on pull requests, including external ones. To avoid double CI
runs, this means we no longer run CI on pushes, except to the master
branch (after a pull request was merged). The deploy job is skipped on
external pull requests, since the secrets won’t be available; this
should cause the browser-tests job to be skipped as well.

Bug: T283911