use example without numeric values
In tables, numbers should be right-eligned in order to make them easy to compare. This is something that implenters could achieve in context. In order to avoid having to make any adjustments to the story, we could use a different example that doesn't include numbers, so we don't display a suboptimal design decision.