Run eslint directly, not through vue-cli
`vue-cli-service lint` (as of recently??) lints only the src/ and tests/
directories [0]. When explicitly targeting the whole directory, it
doesn't apply the `.eslintignore`. It's unclear what the proper way is
to lint all directories, but running eslint directly without the vue-cli
wrapper does the right thing, so let's just do that.

[0]: https://cli.vuejs.org/core-plugins/eslint.html#injected-commands