Add additional build for Vue 3 compatibility mode
This adds an extra build to dist, named wikit-vue-components-vue3compat,
which includes some changes to make Wikit compatible with the Vue 3
migration build:

- the @vue/composition-api dependency is externalized, because it’s
  incompatible with Vue 3
- components that use an HTML element with the same name (e.g.
  Button/<button>) internally use a non-conflicting name (but are still
  imported and used under their original name elsewhere), because
  otherwise there are sometimes infinite loops when mounting them

The intention is that this additional build can be used in two ways:

1. with Vue 2, where 'vue' is vue@^2 and '@vue/composition-api' is
   @vue/composition-api@^1
2. with Vue 3, where 'vue' is @vue/compat@^3 and '@vue/composition-api'
   is an alias for the same module

Note that even in option 1, @vue/composition-api must now be available
in the environment, whereas in the regular (non-vue3compat) build it is
built into the bundle and only Vue itself is required.

Note that this only makes Wikit compatible with the migration build, not
with pure Vue 3: we still use Vue 2 APIs, this just includes the bare
minimum of changes to make the library work under the migration build.