Upgrade storybook to latest version (#532)
This is done in preparation for the migration to Vue 3. Storybook was
previously pinned to 6.0 because of IE11 incompatibilities with 6.1. But
since IE 11 is no longer supported with Vue 3, this is a non-issue now.

The switch to node 16 should mean that we no longer use npm 6. This is intended to fix the follwing strange error:

> bindings not accessible from watchpack-chokidar2:fsevents