Replace stylelint-use-logical with stylelint-use-logical-spec
The original stylelint-use-logical seems to be effectively unmaintained,
and is causing problems with its old stylelint peer dependency; its fork
stylelint-use-logical-spec seems reasonably active, has a comparable
number of downloads, and is compatible with newer stylelint.

The new version also complains about width/height (wanting them replaced
with inline/block-size), but at least one of these replacements is not
trivial (in OptionsMenu, the max-width involves the vw unit), so disable
this for now.