Jump to Vue 3 (#551)
This is a single big commit that migrates all the components in the next
branch to Vue 3.

With the options API, there were a lot of error of typescript not
understanding where props are coming from when they were used in
methods. It is unclear why that is, but it can be fixed by moving the code to
script setup. Therefore that was done for the Button and the Lookup
components.

A further consideration is that Vue 3 now renders comments (at least in
development mode) and so it breaks some tests that expects a single
root.

The tests of the LookupInput in particular needed some more work:

* Vue.nextTick is now nextTick
* disabled is now always the empty string if present
* .native has been removed
** the native focus event has to be triggered on the actual HTML element

Also, many more wrapper methods are now async so that we can just await them
instead of having to wait for nextTick explicitly. This means that
nextTick is no longer needed in the Lookup tests.

A lot of the tests produced confusing typescript errors. The root cause
of these mysterious errors turned out to actually be that we were still
using the propsData key. Apparently that is deprecated when using
javascript tests, but breaks hard when typescript is used.

The tests in LookupCore.spec.ts specifically need also the change to
src/shims-vue.d.ts to work, the others don't.

Also there is a change in Vue3 in how the whitespace coming from
newlines in the template between tags is handled. This affects us
insofar as some of the whitespace we saw in storybook is now gone.
This is fixed by adding a compiler option to those stories. This
generally doesn't affect production because there we are not relying on
that whitespace from newlines in the template, but rather create
whitespace via CSS where we want it.

Also, using props in the storybook stories to work with the Controls
Storybook addon somehow stopped working. So some stories were changed to
inject the args into the `data` option instead.

The comments for the LookupInput events are moving to the same events in
Lookup component, so that they actually show up in storybook.

Typescript also validates that the values given to props match the
excepted types. Though I saw these errors only in my IDE so far, not yet
in the actual type check during npm run build.

Bug: T303503