Cherry Pick "Extract common Icon <svg> attributes (#577)" (#579)
All elements shared five common attributes, which can be extracted
into a common object using v-bind – except that the Search type didn’t
have a fill=none attribute before. I assume this wasn’t intentional (or
at least, there’s no discussion of it in #371), so it should be okay to
add the attribute with this change. (There’s no snapshot to update for
this case because there’s no snapshot specifically for the Search type.)


Co-authored-by: Lucas Werkmeister <lucas.werkmeister@wikimedia.de>