Cherry pick "Wrap button :focus, :hover and :active styles in :enabled (#584)"
Firefox allows a button to be :disabled and still have :focus, if it was
programmatically disabled while focused. In this case, the :focus styles
would beat the :disabled styles, since they have higher specificity; we
want disabled buttons to always look disabled, though, so wrap all the
other styles in :enabled so they only apply when wanted.

The stylelint rule max-nesting-depth (max depth: 3), which complains
about the change, is disabled – the alternative solutions (combining
selectors together) seem worse to us in this case.

Bug: T304511