Make Lookup react better to Enter key events
Make the OptionsMenu prevent default actions for Enter key events it
handles (like it already does for ArrowUp and ArrowDown), but only have
LookupInput send events to OptionsMenu in the first place when the menu
is actually being shown, so that default actions *aren’t* being
prevented when the menu is hidden (e.g. after an option was selected).

Bug: T305359