Cherry pick "Add aria attributes for screen readers (#587)"
Add more specific and complete aria attributes to allow screen
readers to announce menu and menu items correctly.

Note: adding an alert role that is read by screen readers every time the 'no-results' panel is shown is not implemented in this PR. the default behavior is that it is read only once if the text remains the same. This issue will be addressed in a follow-up PR.

Bug: T290733