Update eslint dependencies in docs package
The wikimedia/client config pulls in client-es5, so if we actually want
es6 linting, then we should use wikimedia/client-es6.
That also complains about some shadowed variables, which are fixed in
this commit as well.

This adds eslint-plugin-es as a dev dependency, because somehow eslint
does not manage to find it as a dependency of eslint-config-wikimedia.